task_next_child() s/unsigned int pos/loff_t pos
pos should be loff_t type. It's might be fine to truncate it
to unsigned int in this case, but it's better not to.
E.g. see 9f6e963f06c1 ("proc: fix ->f_pos overflows in first_tid()")

+++ fs/proc/internal.h
@@ -306,9 +306,8 @@ struct tgid_iter {
 };
 struct tgid_iter next_tgid(struct pid_namespace *ns, struct tgid_iter iter);
 
-struct task_struct *
-task_next_child(struct task_struct *parent, struct task_struct *prev, unsigned int pos);
-
+struct task_struct *task_next_child(struct task_struct *parent,
+				struct task_struct *prev, loff_t pos);
 struct task_struct *task_first_tid(struct pid *pid, int tid, loff_t f_pos,
 					struct pid_namespace *ns);
 struct task_struct *task_next_tid(struct task_struct *start);
